refactor: prestring.python._codeobject -> prestring.codeobject
